Rename some subsolver modules.
