+++ src/estimagic/optimization/subsolvers/_conjugate_gradient_fast.py
