+++ src/estimagic/optimization/subsolvers/_steihaug_toint.py
