+++ src/estimagic/optimization/subsolvers/_trsbox.py
