+++ src/estimagic/optimization/subsolvers/_trsbox_fast.py
